+++ R/tide-harmonics.R
@@ -58,14 +58,13 @@ tide_harmonics <- function (x) {
              "units", "longitude", "latitude", "timezone", "tzfile", "datum",
              "A", "kappa") %in% names(x))) stop("x missing components", call. = FALSE)
 
-
-  x$Station <- tibble::data_frame(
+  x$Station <- data.frame(
     Station = x$station, Units = x$unit, Longitude = x$longitude, Latitude = x$latitude,
-    Hours = x$timezone, TZ = x$tzfile, Datum = x$datum)
+    Hours = x$timezone, TZ = x$tzfile, Datum = x$datum, stringsAsFactors = FALSE)
 
   x$Station$Station <- enc2utf8(x$Station$Station)
 
-  x$Node <- tibble::data_frame(Node = x$name, Speed = x$speed)
+  x$Node <- data.frame(Node = x$name, Speed = x$speed, stringsAsFactors = FALSE)
   x$StationNode <- abind::abind(A = x$A, Kappa = x$kappa, along = 3)
   dimnames(x$StationNode) <- list(x$Station$Station, x$Node$Node, c("A", "Kappa"))
 
@@ -83,6 +82,9 @@ tide_harmonics <- function (x) {
   x$Node <- x$Node[node,,drop = FALSE]
   x$StationNode <- x$StationNode[,node,,drop = FALSE]
   x$NodeYear <- x$NodeYear[node,,,drop = FALSE]
+
+  x$Node <- as_conditional_tibble(x$Node)
+  x$Station <- as_conditional_tibble(x$Station)
 
   class(x) <- c("tide_harmonics")
   check_tide_harmonics(x)